+++ contracts/libs/DepositQueueLib.sol
@@ -4,12 +4,12 @@ pragma solidity >=0.8.6;
 library DepositQueueLib {
     struct DepositEntry {
         address owner;
-        uint amount;
+        uint256 amount;
     }
 
     struct DepositQueue {
         address[] list;
-        mapping(address => uint) cache;
+        mapping(address => uint256) cache;
     }
 
     function push(DepositQueue storage queue, DepositEntry memory deposit) external {
@@ -20,7 +20,7 @@ library DepositQueueLib {
         queue.cache[deposit.owner] += deposit.amount;
     }
 
-    function remove(DepositQueue storage queue, uint startIndex, uint endIndex) external {
+    function remove(DepositQueue storage queue, uint256 startIndex, uint256 endIndex) external {
         if (endIndex > startIndex) {
             // Remove the interval from the cache
             while(startIndex < endIndex) {
@@ -30,7 +30,7 @@ library DepositQueueLib {
 
             // Update the list with the remaining entries
             address[] memory newList = new address[](queue.list.length - endIndex);
-            uint i = 0;
+            uint256 i = 0;
 
             while(endIndex < queue.list.length) {
                 newList[i++] = queue.list[endIndex++];
@@ -40,17 +40,17 @@ library DepositQueueLib {
         }
     }
 
-    function get(DepositQueue storage queue, uint index) external view returns(DepositEntry memory depositEntry) {
+    function get(DepositQueue storage queue, uint256 index) external view returns(DepositEntry memory depositEntry) {
         address owner = queue.list[index];
         depositEntry.owner = owner;
         depositEntry.amount = queue.cache[owner];
     }
 
-    function balanceOf(DepositQueue storage queue, address owner) external view returns(uint) {
+    function balanceOf(DepositQueue storage queue, address owner) external view returns(uint256) {
         return queue.cache[owner];
     }
 
-    function size(DepositQueue storage queue) external view returns(uint) {
+    function size(DepositQueue storage queue) external view returns(uint256) {
         return queue.list.length;
     }
 }
