+++ contracts/mocks/Asset.sol
@@ -4,7 +4,7 @@ pragma solidity >=0.8.6;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract Asset is ERC20("Asset", "AST") {
-    function mint(uint amount) public {
+    function mint(uint256 amount) public {
         _mint(msg.sender, amount);
     }
 
@@ -12,7 +12,7 @@ contract Asset is ERC20("Asset", "AST") {
         _mint(to, 10 ether);
     }
 
-    function burn(uint amount) public {
+    function burn(uint256 amount) public {
         _burn(msg.sender, amount);
     }
 }
